Move cleanup execution to sub runner with simulation support
Reverted changes recently made to the 'Check' interface. It wasn't such a bad idea
to pass the 'Result' to a check after all. This makes sense for more complex checks
so the check can make use of the 'debug' functionality.

Finally I decided to go for a check simulator as well.

Issues: #36, #37