Fix simulation bug for uncompressed sources
In order to simulate a phpbu execution file operations like 'is_file', 'is_dir'
or 'file_exists' can not be executed in simulation mode.
So to eliminate this calls in simulation mode I moved some file validation
into source classes and added a 'isDirectory' method to the  source 'Status'.

Status::uncompressed is now deprecated and will be removed in phpbu 4.0
One of Status::uncompressedFile or Status::uncompressedDirectory should
be used instead.