Add rsync source
The 'rsync' source enables you to collect files from a remote server. For example,
if your backup server is located behind a firewall the backup server has to pull the backups
instead of all your servers pushing their backups to it.
This is now possible by running phpbu on your backup server and using the 'rsync' source.