Change Str::toBoolean method to accept boolean
If you call the toBoolean method with a boolean instead of a string
it will directly return the boolean without any changes.
This has some impact on the json configuration files. Now it is
possible to use booleans for boolean values like "lockTables".
So you don't have to use bool-strings like "true" or "false" any more.
You still can, but I think the more intuitive way is to use boolean values.