Add configuration adapter handling
I decided to use adapters to manipulate the configuration values
before creating the actual configuration. So the application does
not have to know anything about configuration adapters.

But to actually create adapters while handling the configuration,
I had to change the configuration loader interface in a backwards
incompatible way. So we have to increase the major version.