Fix 'deleteTarget' simulation problem - issue #113
There was an error if the 'deleteTarget' option was activated
during simulation. The target is getting added to the files
to delete list and the Cleaner was trying to get the MTime
and Size from a file that doesn't exist.
To fix this we deactivate the option during simulation and
add some hints to the output, that the target will be deleted
as well.