Upgrade to 'cli' lib version 2.0
With version 2 of the cli lib we can define what kind of exit
codes we are willing to accept without any exceptions.
This was necessary to handle the exit codes of 'tar' in case
of failed read warnings.

Now, in case of 'ignoreFailedRead' option to be true, we accept
exit code 1 from tar as well.

Issue #114