Handle bootstrap before config creation - issue #124
Bootstrapping and config creation is kind of a chicken and egg
problem. To actually use a custom adapter the bootstrap file has
to be loaded before creating the configuration.

So from now on all config loader have to use the new Bootstrapper
to handle the bootstrap loading before creating the actual
configuration instance.