Fix local collector issues
Collectors did only work if last path element was a
date placeholder. There is a unit test to demonstrate
show that behaviour (testMixedStaticEnd).

Since the target path is always of a fixed depth the
new strategy is to always check against this depth to
make sure we are looking for files in a path equally as
deep as the target path.