Fix target timing issue #156
In cases where sources create a directory and phpbu has to compress the
directory with tar first phpbu will add a '.tar' suffix to the target.

The regex to collect all backups for cleanup was created before all
suffixes where set correctly. This resulted in not working cleanups.

The issue is fixed by creating the regex right before all backups
are collected.