Activate incremental backup support
The tar source can now handle incremental backups
if supplied with a metadata file to cache the current
state.

You can use only the 'incrementalFile' option. That
leaves you in charge of handling the backup level
and snar file handling.

If you do not want to handle it yourself you can supply
the 'forceLevelZeroOn' option. This will make sure at
configured points in time a level zero backup is
executed

This fixes issue #105