Fix debug-flag in configuration files (#323)
According to the documentation, both the XML and JSON configuration
format should support setting a `"debug"` flag. However, the XSD
schema does not allow a `debug` attribute on the root `<phpbu>` node.
Similarly, the JSON configuration loader ignores a top-level `"debug"`
key, even though one is present in the example configuration in
`\phpbu\App\Configuration\Loader\Json`.

This commit amends the XSD schma with an optional debug attribute on
the phpbu-Type. Additionally, support for loading this value is added
to the XML and JSON loader.